Remove RVD and UnpartitionedRVD (#4385)
* remove UnpartitionedRVD

* delete UnpartitionedRVD.scala
* change calls to `UnpartitionedRVD.apply` to `OrderedRVD.unkeyed`

* Remove uses of RVD

* Replace uses of `RVD` by `OrderedRVD`
* Remove `RVD` overload of `OrderedRVD.coerce`, replace uses by `OrderedRVD.changeKey`
* Remove `toOldStyleRVD` and `toOrderedRVD`, which are now no-ops

* Move method implementations from RVD to OrderedRVD

* actually delete RVD trait

* Remove RVD object

* Remove `RVD` object. It’s methods are scattered:
  * `writeLocalUnpartitioned` -> `RVDSpec.writeLocal`
  * `union` folded into `OrderedRVD.union`
  * `memoryCodec` and `wireCodec` moved to `OrderedRVD`
  * `regionValueToBytes` -> `RegionValue.toBytes`
  * `bytesToRegionValue` -> `RegionValue.fromBytes`
* Refactor `stabilize`, `destabilize`, etc.

* remove redundant casts to OrderedRVD

* fix bgen

* remove unnecessary scan from fromRowsTable

* force retest
