+++ hail/src/main/scala/is/hail/annotations/RegionValue.scala
@@ -1,11 +1,10 @@
 package is.hail.annotations
 
-import java.io.{ByteArrayInputStream, ByteArrayOutputStream, ObjectInputStream, ObjectOutputStream}
+import java.io._
 
-import com.esotericsoftware.kryo.io.{Input, Output}
-import com.esotericsoftware.kryo.{Kryo, KryoSerializable}
 import is.hail.expr.types.Type
-import is.hail.io.{CodecSpec, Decoder, Encoder}
+import is.hail.utils.using
+import is.hail.io.{Decoder, Encoder}
 import sun.reflect.generics.reflectiveObjects.NotImplementedException
 
 object RegionValue {
@@ -14,6 +13,19 @@ object RegionValue {
   def apply(region: Region): RegionValue = new RegionValue(region, 0)
 
   def apply(region: Region, offset: Long) = new RegionValue(region, offset)
+
+  def fromBytes(
+    makeDec: InputStream => Decoder,
+    r: Region,
+    carrierRv: RegionValue
+  )(bytes: Array[Byte]
+  ): RegionValue =
+    using(new ByteArrayInputStream(bytes)) { bais =>
+      using(makeDec(bais)) { dec =>
+        carrierRv.setOffset(dec.readRegionValue(r))
+        carrierRv
+      }
+    }
 }
 
 final class RegionValue(
@@ -34,6 +46,15 @@ final class RegionValue(
   }
 
   def pretty(t: Type): String = region.pretty(t, offset)
+
+  def toBytes(makeEnc: OutputStream => Encoder): Array[Byte] =
+    using(new ByteArrayOutputStream()) { baos =>
+      using(makeEnc(baos)) { enc =>
+        enc.writeRegionValue(region, offset)
+        enc.flush()
+        baos.toByteArray
+      }
+    }
 
   private def writeObject(s: ObjectOutputStream): Unit = {
     throw new NotImplementedException()
