+++ hail/src/main/scala/is/hail/expr/ir/MatrixIR.scala
@@ -258,7 +258,7 @@ case class MatrixNativeReader(path: String) extends MatrixReader {
         val fullRowType = requestedType.rvRowType
         val localEntriesIndex = requestedType.entriesIdx
 
-        val rowsRVD = spec.rowsComponent.read(hc, path, requestedType.rowType).asInstanceOf[OrderedRVD]
+        val rowsRVD = spec.rowsComponent.read(hc, path, requestedType.rowType)
         if (mr.dropCols) {
           val (t2, makeF) = ir.Compile[Long, Long](
             "row", requestedType.rowType,
@@ -1320,9 +1320,7 @@ case class MatrixMapRows(child: MatrixIR, newRow: IR, newKey: Option[(IndexedSeq
     }
 
     val newRVD = if (newKey.isDefined) {
-      OrderedRVD.coerce(
-        typ.orvdType,
-        prev.rvd.mapPartitionsWithIndex(typ.rvRowType, mapPartitionF))
+      prev.rvd.mapPartitionsWithIndex(typ.rvRowType, mapPartitionF).changeKey(typ.rowKey)
     } else {
       prev.rvd.mapPartitionsWithIndexPreservesPartitioning(typ.orvdType, mapPartitionF)
     }
@@ -1883,7 +1881,6 @@ case class MatrixAnnotateRowsTable(
               ) ++ newKeyUIDs.zip(newKeys)),
             None))).execute(hc)
         val indexedRVD1 = mrt.rvd
-          .asInstanceOf[OrderedRVD]
           .zipWithIndex(indexUID, Some(partitionCounts))
         val tl1 = TableLiteral(mrt.copy(typ = mrt.typ.copy(rowType = indexedRVD1.rowType), rvd = indexedRVD1))
 
@@ -1898,8 +1895,8 @@ case class MatrixAnnotateRowsTable(
                   .reduce[IR] { case(l, r) => ApplySpecial("||", FastIndexedSeq(l, r))})),
             newKeyUIDs)).execute(hc)
 
-        val left = sortedTL.rvd.asInstanceOf[OrderedRVD]
-        val right = tv.rvd.asInstanceOf[OrderedRVD]
+        val left = sortedTL.rvd
+        val right = tv.rvd
         val joined = left.orderedLeftJoinDistinctAndInsert(right, root)
         val prevPartitioner = prev.rvd.partitioner
 
@@ -1908,7 +1905,7 @@ case class MatrixAnnotateRowsTable(
         val indexedPartitioner = prevPartitioner.copy(
           kType = TStruct((prevRowKeys ++ Array(indexUID)).map(fieldName => fieldName -> joined.typ.rowType.field(fieldName).typ): _*))
         val oType = joined.typ.copy(key = prevRowKeys ++ Array(indexUID))
-        val rpJoined = OrderedRVD.shuffle(oType, indexedPartitioner, joined)
+        val rpJoined = OrderedRVD.shuffle(oType, indexedPartitioner, joined.crdd)
 
         val indexedMtRVD = prev.rvd.zipWithIndex(indexUID, Some(partitionCounts))
 
@@ -1923,7 +1920,7 @@ case class MatrixAnnotateRowsTable(
       case None =>
         assert(child.typ.rowKeyStruct.types.zip(table.typ.keyType.types).forall { case (l, r) => l.isOfType(r) })
         val newRVD = prev.rvd.orderedLeftJoinDistinctAndInsert(
-          tv.rvd.asInstanceOf[OrderedRVD], root)
+          tv.rvd, root)
         prev.copy(typ = typ, rvd = newRVD)
     }
   }
@@ -2046,7 +2043,7 @@ case class TableToMatrixTable(
 
     val ordType = OrderedRVDType(rowKey ++ FastIndexedSeq(INDEX_UID), rowEntryStruct)
     val ordTypeNoIndex = OrderedRVDType(rowKey, rowEntryStruct)
-    val ordered = OrderedRVD.coerce(ordType, rowKey.length, rowEntryRVD)
+    val ordered = rowEntryRVD.changeKey(ordType.key, rowKey.length)
     val orderedEntryIndices = entryFields.map(rowEntryStruct.fieldIdx)
     val orderedRowIndices = (rowKey ++ rowFields).map(rowEntryStruct.fieldIdx)
 
@@ -2362,7 +2359,7 @@ case class UnlocalizeEntries(rowsEntries: TableIR, cols: TableIR, entryFieldName
     val (_, missingF) = ir.Compile[Long, Boolean]("row", rowsEntries.typ.rowType,
       ir.IsNA(field))
 
-    var rowOrvd = rowtab.rvd.asInstanceOf[OrderedRVD]
+    var rowOrvd = rowtab.rvd
     rowOrvd = rowOrvd.mapPartitionsWithIndexPreservesPartitioning(rowOrvd.typ) { (i, it) =>
       val missing = missingF(i)
       val len = lenF(i)
