+++ hail/src/test/scala/is/hail/expr/ir/TableIRSuite.scala
@@ -217,7 +217,7 @@ class TableIRSuite extends SparkSuite {
         Some(1)),
       if (!leftProject.contains(1)) IndexedSeq("A", "B") else IndexedSeq("A")))
     val partitionedLeft = left.copy2(
-      rvd = left.value.rvd.asInstanceOf[OrderedRVD]
+      rvd = left.value.rvd
         .constrainToOrderedPartitioner(if (!leftProject.contains(1)) leftPart else leftPart.coarsen(1)))
 
     val (rightType, rightProjectF) = rowType.filter(f => !rightProject.contains(f.index))
@@ -227,7 +227,7 @@ class TableIRSuite extends SparkSuite {
         Some(1)),
       if (!rightProject.contains(1)) IndexedSeq("A", "B") else IndexedSeq("A")))
     val partitionedRight = right.copy2(
-      rvd = right.value.rvd.asInstanceOf[OrderedRVD]
+      rvd = right.value.rvd
         .constrainToOrderedPartitioner(if (!rightProject.contains(1)) rightPart else rightPart.coarsen(1)))
 
     val (_, joinProjectF) = joinedType.filter(f => !leftProject.contains(f.index) && !rightProject.contains(f.index - 2))
